services/chatbot/requirements.txt
# ============================================================================
# ChatBot Service - AI Assistant Project
# Python: 3.10.6
# Description: Multi-model chatbot with local LLM support and image generation
# ============================================================================

# ============= WEB FRAMEWORK =============
flask==3.1.3
flask-cors==6.0.2  # Updated from 4.0.0 - Security fix (5 vulnerabilities patched)
python-dotenv==1.2.2

# ============= FASTAPI (async mode) =============
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9  # Required for UploadFile
pydantic>=2.0.0
werkzeug>=3.0.0

# ============= PERFORMANCE & CACHING (Phase 1) =============
redis>=5.0.0
pymongo>=4.6.0
dnspython>=2.4.0  # Required for MongoDB SRV connection

# ============= RAG =============
chromadb>=0.5.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
pgvector>=0.3.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
beautifulsoup4>=4.12.0
pymupdf>=1.24.0

# ============= AI APIs =============
openai>=2.29.0  # Videos API (Sora 2) requires >=2.29
google-genai>=1.56.0

# ============= IMAGE PROCESSING =============
Pillow>=10.0.0
requests>=2.31.0

# ============= PHASE 2: ADVANCED FEATURES =============
# Semantic search and embeddings
sentence-transformers>=2.2.2

# Scientific computing
scipy>=1.11.0
scikit-learn>=1.3.0

# ============= LOCAL MODEL SUPPORT =============
# PyTorch (CUDA 11.8) - Install from https://pytorch.org/get-started/locally/
numpy<3.0.0  # Required for PyTorch compatibility
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.37.0
accelerate>=0.27.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Quantization (for efficient local models - Windows requires Visual C++)
bitsandbytes>=0.41.0; sys_platform != 'win32'

# ============= ADDITIONAL UTILITIES =============
# Progress bars
tqdm>=4.65.0

# JSON handling
jsonschema>=4.17.0

# Safe loading
pyyaml>=6.0

# File handling
aiofiles>=23.0.0

# ============= GOOGLE CLOUD / DRIVE =============
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.108.0
firebase-admin>=7.3.0